feat(app): add search results and primary search react components
