+++ client/components/layout/header/index.tsx
@@ -6,7 +6,7 @@ import { Logo } from '@/components/ui/logo'
 import { HeaderIcons } from './header-icons'
 import { Navigation } from '../navigation'
 import { AuthDropdown } from '@/components/features/user/auth-dropdown'
-import { Search } from '@/components/features/search'
+import { Search } from '@/components/features/full-search'
 import { useCurrentUser } from '@/hooks/useCurrentUser'
 import { useAuthContext } from '@/context/auth-context'
 import { HeaderBasic } from './header-basic'
